Rename RunDetailsV4 as RunDetails
