+++ ui/packages/components/src/RunDetails/ErrorInfo.tsx
