+++ ui/packages/components/src/RunDetails/utils/formatting.ts
