+++ ui/packages/components/src/RunDetails/utils/traceConversion.ts
@@ -174,9 +174,7 @@ function tracesToBarData(
     if (bar.endTime) {
       const endMs = bar.endTime.getTime();
       latestCompletedSiblingEndMs =
-        latestCompletedSiblingEndMs === null
-          ? endMs
-          : Math.max(latestCompletedSiblingEndMs, endMs);
+        latestCompletedSiblingEndMs === null ? endMs : Math.max(latestCompletedSiblingEndMs, endMs);
     }
 
     bars.push(bar);
@@ -246,8 +244,8 @@ function traceToBarData(
   const traceStartedAtMs = trace.startedAt ? new Date(trace.startedAt).getTime() : null;
   const traceQueuedAtMs = trace.queuedAt ? new Date(trace.queuedAt).getTime() : null;
   const childDiscoveryStartAtMs = trace.isRoot
-    ? (traceStartedAtMs ?? discoveryStartAtMs)
-    : (traceStartedAtMs ?? traceQueuedAtMs ?? discoveryStartAtMs);
+    ? traceStartedAtMs ?? discoveryStartAtMs
+    : traceStartedAtMs ?? traceQueuedAtMs ?? discoveryStartAtMs;
 
   // Check if this step has experiment metadata
   const hasExperiment = trace.metadata?.some((m) => m.kind === KindInngestExperiment) ?? false;
